# gin-blog-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/code": {
            "get": {
                "description": "发送邮箱验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserAuth"
                ],
                "summary": "发送邮箱验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/comment": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据条件查询评论列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "条件查询评论列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "昵称",
                        "name": "nickname",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "审核状态",
                        "name": "is_review",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "评论类型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page_num",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-model_CommentVO"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据 ID 数组删除评论",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "删除评论（批量）",
                "parameters": [
                    {
                        "description": "评论 ID 数组",
                        "name": "ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-int"
                        }
                    }
                }
            }
        },
        "/comment/review": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据 ID 数组修改审核状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "修改评论审核（批量）",
                "parameters": [
                    {
                        "description": "修改审核状态",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateReviewReq"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-any"
                        }
                    }
                }
            }
        },
...
        "/link": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "添加或修改友链",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Link"
                ],
                "summary": "添加或修改友链",
                "parameters": [
                    {
                        "description": "添加或修改友链",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AddOrEditLinkReq"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-model_FriendLink"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据 ID 数组删除友链",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Link"
                ],
                "summary": "删除友链（批量）",
                "parameters": [
                    {
                        "description": "友链ID数组",
                        "name": "ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-int64"
                        }
                    }
                }
            }
        },
        "/link/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据条件查询获取友链列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Link"
                ],
                "summary": "获取友链列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "当前页数",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "page_num",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键字",
                        "name": "keyword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-handler_PageResult-model_FriendLink"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserAuth"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "登录",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-model_LoginVO"
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "description": "注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserAuth"
                ],
                "summary": "注册",
                "parameters": [
                    {
                        "description": "注册",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.RegisterReq"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.AddOrEditLinkReq": {
            "type": "object",
            "required": [
                "address",
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "intro": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "handler.LoginReq": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "handler.PageResult-model_FriendLink": {
            "type": "object",
            "properties": {
                "page_data": {
                    "description": "分页数据",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FriendLink"
                    }
                },
                "page_num": {
                    "description": "每页条数",
                    "type": "integer"
                },
                "page_size": {
                    "description": "上次页数",
                    "type": "integer"
                },
                "total": {
                    "description": "总条数",
                    "type": "integer"
                }
            }
        },
...
        "handler.RegisterReq": {
            "type": "object",
            "required": [
                "code",
                "password",
                "username"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 4
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "handler.Response-any": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
        "handler.Response-handler_PageResult-model_FriendLink": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/handler.PageResult-model_FriendLink"
                        }
                    ]
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
        "handler.Response-int": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据",
                    "type": "integer"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
        "handler.Response-int64": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据",
                    "type": "integer"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
        "handler.Response-model_CommentVO": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.CommentVO"
                        }
                    ]
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
        "handler.Response-model_FriendLink": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.FriendLink"
                        }
                    ]
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
        "handler.Response-model_LoginVO": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.LoginVO"
                        }
                    ]
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
        "handler.UpdateReviewReq": {
            "type": "object",
            "required": [
                "is_review"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "is_review": {
                    "type": "integer",
                    "maximum": 1,
                    "minimum": 0
                }
            }
        },
...
        "model.CommentVO": {
            "type": "object",
            "properties": {
                "article_title": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_review": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "reply_nickname": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
        "model.FriendLink": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "intro": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.LoginVO": {
            "type": "object",
            "properties": {
                "article_like_set": {
                    "description": "点赞 Set: 用于记录用户点赞过的文章, 评论",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "avatar": {
                    "type": "string"
                },
                "comment_like_set": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "intro": {
                    "type": "string"
                },
                "ip_address": {
                    "type": "string"
                },
                "ip_source": {
                    "type": "string"
                },
                "last_login_time": {
                    "type": "string"
                },
                "login_type": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "user_info_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
...
    }
}